onboarding/requirements.txt
async-timeout==4.0.3
asyncpg==0.29.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
soupsieve==2.5
